feat: [Setting] redux 설정
