--- src/App.js
@@ -1,24 +1,0 @@
-import React from "react";
-import { BrowserRouter, Routes, Route } from "react-router-dom";
-import "./App.css";
-import Main from "./pages/Main";
-import User from "features/user/User";
-import Diary from "features/diary/Diary";
-import Page from "features/page/Page";
-
-function App() {
-  return (
-    <div className="App">
-      <BrowserRouter>
-        <Routes>
-          <Route path="/" element={<Main />} />
-          <Route path="/user" element={<User />} />
-          <Route path="/diary" element={<Diary />} />
-          <Route path="/page" element={<Page />} />
-        </Routes>
-      </BrowserRouter>
-    </div>
-  );
-}
-
-export default App;
