+++ src/app/store.ts
@@ -1,0 +1,14 @@
+import { configureStore } from "@reduxjs/toolkit";
+import { combineReducers } from "redux";
+
+const rootReducer = combineReducers({
+  // 설정한 reducer
+});
+
+const store = configureStore({
+  reducer: rootReducer,
+});
+
+export default store;
+export type AppDispatch = typeof store.dispatch;
+export type rootState = ReturnType<typeof store.getState>;
